Added metadata generation

+++ YoutubeYownloader.py
@@ -8,12 +8,14 @@ from PIL import Image
 
 import subprocess
 import os
+import os.path
 import re
 import urllib
 
 
 url = input("URL:")
 
+#switches from webm to mp3
 def fix_audio(path):
     subprocess.run([
         "ffmpeg",
@@ -25,14 +27,19 @@ def fix_audio(path):
               path + ".mp3"
     ], stdout=subprocess.DEVNULL, stderr=subprocess.DEVNULL)
 
+#formats title properly, removing undesirables
 def fix_title(title):
-    nonowords=["Kasane Teto", "feat."]
+    nonowords=["Kasane Teto", "feat.", "?", "\""]
 
     for word in nonowords:
         title = re.sub(re.escape(word),"",title)
 
+
     title = re.sub(r'[\./]',"",title)
+
+    #uncomment to remove all parenthetical expressions
     #title = re.sub(r'\(.*\)', "", title)
+
     title = re.sub(r'\s*$', "", title)
     title = re.sub(r'^\s*', "", title)
 
@@ -40,6 +47,7 @@ def fix_title(title):
 
     return title
 
+#downloads video thumbnail to "thumbnail.jpg"
 def download_thumbnail(video):
     try:
         thumbnail_url = video.thumbnail_url
@@ -49,6 +57,7 @@ def download_thumbnail(video):
     except Exception as e:
         print(f"An error occurred: {e}")
 
+#crops current "thumbnail.jpg to largest possible square
 def crop_thumbnail():
     #open image
     img = Image.open("thumbnail.jpg")
@@ -68,7 +77,8 @@ def crop_thumbnail():
     #save
     square_img.save("thumbnail.jpg", "JPEG")
 
-def add_metadata(audio_file_path, video, playlist=None):
+#adds metadata to the mp3 file, this header is bloated and bad, do not do this
+def add_metadata(audio_file_path, video, playlist=None, track_number=None):
     #check exists
     if not os.path.exists(audio_file_path):
         print(f"Error: {audio_file_path} not found.")
@@ -82,10 +92,13 @@ def add_metadata(audio_file_path, video, playlist=None):
             audio.add_tags()
 
         #set metadata
-        audio['title'] = video.title
+        audio['title'] = fix_title(video.title)
         audio['artist'] = re.sub(r' - Topic',"",video.author)
         if playlist:
-            audio['album'] = playlist
+            audio['album'] = "Indie Game: The Movie"
+        if track_number:
+            audio['tracknumber'] = track_number
+
 
         #save
         audio.save()
@@ -107,23 +120,29 @@ def add_metadata(audio_file_path, video, playlist=None):
             )
         audio_full.save()
 
-def process_video(path,video, playlist=None):
+#processes one video, this header is bloated and bad, do not do this
+def process_video(path,video, playlist=None, track_number=None):
     title= "Title Grab Failed"
 
     try:
         audio = video.streams.get_audio_only()
 
         title=fix_title(video.title)
+
         print(title)
+
+        #skip existing songs
+        if os.path.isfile(path + "\\" + f"{title}.mp3"):
+            return
+
         audio.download(output_path=path, filename=f"{title}.webm")
         fix_audio(path + "\\" + f"{title}")
         os.remove(path + "\\" + f"{title}.webm")
 
-        add_metadata(path + "\\" + f"{title}.mp3", video,playlist)
+        add_metadata(path + "\\" + f"{title}.mp3", video,playlist, track_number)
     except None:
         print("Failed: "+title)
         print("Continuing")
-
 
 
 if "playlist" in url:
@@ -134,12 +153,12 @@ if "playlist" in url:
     print("Playlist: "+playlist.title)
     print("Playlist Size: "+str(len(videos)))
 
+    i=1
     for v in videos:
-
-        process_video(playlist.title,v,playlist.title)
+        process_video("Indie Game The Movie",v,playlist.title, str(i)+'/'+str(len(videos)))
+        i+=1
 
 else:
-
     video = YouTube(url)
     process_video("output", video)
